requirements.txt
python-dotenv
pandas
numpy
pymongo
certifi
pymongo[srv]
scikit-learn
dill
pyyaml
xgboost
fastapi
uvicorn
ipykernel
apache-airflow
mlflow
dvc
jinja2
python-multipart
# -e refers to setup.py and . refers to current directory
ensure==1.0.4
python-box==7.3.2
joblib
optuna
catboost
-e .